need to restructure the way weapon damage is done
currently the weapon damage is done when card clicked on, assigning an
attacking player and a defending one, then when the card is left inside
the playarea, and the current defenders body parts are collected to see
if they are targetted, then given damage.
Two problems with this. The imediate one, with a delay between
targetting the part and clicking to damage it, and then actually being
damaged, cannot be broken apart without leaving the card and targetting
squares active for that delay. If there is going to be a targettted
strike, followed by a delay in which the opponent can react, then have
damage taken, the info of the attack needs to be stores.
Second, the process is completely vunerable to overlap if the opponent
is going to be doing actions at the same time as the player, as the card
clicking and card releasing are so far apart.
Also, organization of the damage is horrible and not intuitive in any
way. Needs to be streamlined and much more consice