fixed the weapon damage delay
moved weapon damage from the game controller down to the player and
playarea controllers to deal with the delay of the weapon damage.

Works well, slight fault in the 'click on small square', unable to
differentaite the click of the square if the card is in play. Would have
to have two mouse inputs for it to fault out, so leaving as is for now.