body part picking flow fixed
much reorganization of the flow of the ui of the body picking settup,
all so the picked modules can be attached to the body parts that store
them. New class created to store body parts id's along with that of the
modules.

need to work on storing them with the body part before verifying and
sending the info to the next screen. At the moment the modules are just
stored in an array.