disabling cards graphic from broken limbs almost
The graphic displays when broken. Problem with duplicate limbs though,
the selection process is only made for 5 limbs, not 8 when including
arms shoulders and legs. Need to go through and correct them all to
include left and right.
Also deselection of modules is currently sending through a -1 and the
selection process isn't weeding them out and counting as a deselect.